+++ lib/pbmenv/version_pathname.rb
@@ -1,0 +1,45 @@
+module Pbmenv
+  class VersionPathname
+    PBM_DIR = "/usr/share/pbm"
+
+    def initialize(version)
+      @version = version
+    end
+
+    def version_path
+      File.join(PBM_DIR, "/v#{@version}")
+    end
+
+    def version_path_without_v
+      File.join(PBM_DIR, "/#{@version}")
+    end
+
+    def app_rb_path
+      File.join(version_path, "app.rb")
+    end
+
+    def app_rb_erb_path
+      File.join(version_path, "app.rb.erb")
+    end
+
+    def device_id_path_in_version
+      File.join(version_path, "/device_id")
+    end
+
+    def device_id_path_in_shared
+      File.join(self.class.shared, "/device_id")
+    end
+
+    def self.device_id_path_in_shared
+      File.join(shared, "/device_id")
+    end
+
+    def self.current
+      File.join(PBM_DIR, "/current")
+    end
+
+    def self.shared
+      File.join(PBM_DIR, "/shared")
+    end
+  end
+end